% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airfoil.R
\docType{data}
\name{airfoil}
\alias{airfoil}
\title{Airfoil data}
\format{
An object of class \code{"data.frame"}.
}
\source{
The UCI Archive https://archive.ics.uci.edu/ml/datasets/Airfoil+Self-Noise,
}
\usage{
data(airfoil)
}
\description{
Here goes a description of the data.
}
\details{
Here goes a more detailed description of the data.
There are 1503 observations and 6 variables:
\code{y}, \code{frequency}, \code{angle}, \code{chord_length},
\code{velocity}, and \code{thickness}.
}
\examples{
data(airfoil)
}
\references{
Brooks, T. F., Pope, D. S., and Marcolini, M. A. (1989). Airfoil self-noise and prediction. NASA Reference Publication-1218, document id: 9890016302.
}
\keyword{datasets}
