% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boost.R
\name{Boost.control}
\alias{Boost.control}
\title{Boost.control}
\usage{
Boost.control(n_init = 100, shrinkage = 1, cc_s = NULL,
  eff_m = NULL, bb = 0.5, trim_prop = NULL, trim_c = 3,
  max_depth_init = 3, min_leaf_size_init = 10, cal_imp = TRUE,
  save_f = FALSE, make_prediction = TRUE, save_tree = FALSE)
}
\arguments{
\item{n_init}{number of iterations for the 1st stage of RRBoost ($T_{1,max}$) (int)}

\item{cc_s}{tuning constant of tukey's loss in SBoost (numeric)}

\item{bb}{breakdown point of the SBoost estimator (numeric)
one per explanatory variable.}

\item{trim_prop}{trimming proportion if `trmse` is used as the performance metric (numeric)}

\item{trim_c}{the trimming constant if `trmse` is used as the performance metric (numeric)}

\item{max_depth_init}{the maximum depth of the initial LADTtree  (numeric, default 3)}

\item{min_leaf_size_init}{the minimum number of observations per node of the initial LADTtree (numeric)}

\item{cal_imp}{calculate variable importance  (TRUE or FALSE)}

\item{save_f}{save the function estimates at all iterations (TRUE or FALSE)}

\item{make_prediction}{make predictions using x_test  (TRUE or FALSE)}

\item{save_tree}{save trees at all iterations  (TRUE or FALSE)}

\item{eff_s}{normal efficiency of tukey's loss in RRBoost (2nd stage) (numeric)}
}
\value{
A list of all input parameters
}
\description{
Control for Boost fits
}
\details{
Various parameters that control aspects of the `Boost` fit.
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
