% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boost.R
\name{Boost.validation}
\alias{Boost.validation}
\title{Boost.validation}
\usage{
Boost.validation(x_train, y_train, x_val, y_val, x_test, y_test,
  type = "RRBoost", error = c("rmse", "aad"), niter = 1000,
  max_depth = 1, y_init = "LADTree", max_depth_init_set = c(1, 2, 3,
  4), min_leaf_size_init_set = c(10, 20, 30),
  control = Boost.control())
}
\arguments{
\item{x_train}{predictor matrix for training data (matrix/dataframe)}

\item{y_train}{response vector for training data (vector/dataframe)}

\item{x_val}{predictor matrix for validation data (matrix/dataframe)}

\item{x_test}{predictor matrix for test data (matrix/dataframe, optional, required when make_prediction in control = TRUE)}

\item{y_test}{response vector for test data (vector/dataframe,  optional, required when make_prediction in control = TRUE)}

\item{error}{types of the error metric on the test set: "rmse","aad"(average absulute deviation), or "trmse" (trimmed rmse) (array)}

\item{niter}{number of iterations (for RRBoost T_{1,max} + T_{2,max}) (numeric)}

\item{max_depth}{the maximum depth of the tree learners (numeric)}

\item{y_init}{the initial estimator, "median" or "LADTree" (string)}

\item{max_depth_init_set}{a vector of possible values of the maximum depth of the initial LADTree that the algorithm choses from}

\item{min_leaf_size_init_set}{a vector of possible values of the minimum observations per node of the initial LADTree that the algorithm choses from}

\item{control}{control parameters specified with Boost.control()}

\item{y_train}{response vector for validation data (vector/dataframe)}
}
\value{
A list with components
\item{the components of model}{an object returned by Boost that is trained with selected initialization parameters}
\item{param}{a vector of selected initialization parameters (return (0,0) if selected initialization is the median of the training responses)}
}
\description{
A function to fit RRBoost where the initialization parameters are chosen by validation
}
\details{
A function to fit RRBoost where the initialization parameters are chosen by validation
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
