% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boost.R
\name{cal_imp}
\alias{cal_imp}
\title{cal_imp}
\usage{
cal_imp(model, x_train, y_train)
}
\arguments{
\item{model}{an object returned by Boost}

\item{x_train}{predictor matrix for test data (matrix/dataframe)}

\item{y_train}{response vector for test data (vector/dataframe)}
}
\value{
The input model list with an additional component of
\item{var_importance}{a vector of permutation variable importance}
}
\description{
A function to calculate variable importance given an object returned by Boost and training data
}
\details{
A function to calculate variable importance given an object returned by Boost and training data
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
