% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boost.R
\name{cal_imp_func}
\alias{cal_imp_func}
\title{cal_imp_func}
\usage{
cal_imp_func(model, x_val, y_val)
}
\arguments{
\item{model}{an object returned by Boost}

\item{x_train}{predictor matrix for test data (matrix/dataframe)}

\item{y_train}{response vector for test data (vector/dataframe)}
}
\value{
\item{var_importance}{a vector of permutation variable importance}
}
\description{
A function to calculate variable importance given an object returned by Boost and validation data
}
\details{
A function to calculate variable importance given an object returned by Boost and validation data
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
