% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boost.R
\name{cal_predict}
\alias{cal_predict}
\title{cal_predict}
\usage{
cal_predict(model, x_test, y_test)
}
\arguments{
\item{model}{an object returned by Boost}

\item{x_test}{predictor matrix for test data (matrix/dataframe)}

\item{y_test}{response vector for test data (vector/dataframe)}
}
\value{
The input model list with the following additional components:

\item{f_t_test}{predicted values with model using x_test as the predictors}
\item{err_test}{a matrix of test errors (returned if make_prediction = TRUE in control)}
\item{f_test}{matrix of test function estimates at all iterations (returned if save_f = TRUE in control)}
\item{value}{a vector of test error evaluated at early stopping time}
}
\description{
A function to make predictions and calculate test error given an object returned by Boost and test data
}
\details{
A function to make predictions and calculate test error given an object returned by Boost and test data
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
